Minor trainer refactor
