+++ ml-agents/mlagents/trainers/components/reward_signals/reward_signal_factory.py
@@ -10,9 +10,6 @@ from mlagents.trainers.components.reward_signals.gail.signal import GAILRewardSi
 from mlagents.trainers.components.reward_signals.curiosity.signal import (
     CuriosityRewardSignal,
 )
-from mlagents.trainers.components.reward_signals.gail.signal import (
-    GAILRewardSignal,
-)
 from mlagents.trainers.tf_policy import TFPolicy
 
 logger = logging.getLogger("mlagents.trainers")
